+++ packages/environment-check/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @sap-ux/environment-check
 
+## 1.0.27
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/project-access 2.1.13 → 2.2.0
+- @sap-ux/axios-extension 2.0.10 → 2.0.10
+
+## 1.0.26
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/axios-extension 2.0.9 → 2.0.10
+
 ## 1.0.25
 
 ### Patch Changes